Remove lru cache and change to use the greedy dual by default
